combine create quiz with create question

+++ app/src/main/java/com/cmloopy/quizzi/data/api/QuestionCreate/QuestionAPI.java
@@ -10,7 +10,7 @@ import retrofit2.http.Path;
 
 public interface QuestionAPI {
     @GET("questions/quiz/{quizId}")
-    Call<List<Question>> getQuizQuestions(@Path("quizId") int quizId);
+    Call<List<Question>> getQuizQuestions(@Path("quizId") Long quizId);
     @GET("questions/{questionId}")
     Call<Question> getQuestionById(@Path("questionId") int questionId);
